refactor: 💡 make ts
